Check for presence of COM object (#19)
This fixes a lot of problems with a lack of checking return types in the various methods that create borehole, log, depth, header, ODBC, title and workspace COM objects. Really it shouldn't be relied on, however.

Individual COM calls that could return `None` or different types should be handled at the point they are called. This PR is just a bit of "backup" so when you try to construct a `Borehole`, `Log`, `Depth`, `Header`, `Odbc`, `Title` or `Workspace` by passing in None, it fails.

Reviewed-on: https://git.alt.lu/alt/pywellcad/pulls/19
Co-authored-by: Matthew Schubert <matthew.schubert@alt.lu>
Co-committed-by: Matthew Schubert <matthew.schubert@alt.lu>