Emulate COM error handling by propagating None (#24)
COM errors seem to be indicated by returning `NULL` (or `None` in Python) rather than a COM object. PR #19 added some error handling for this which would throw an exception if a class tried to wrap a `None` object.

This PR changes that behaviour and instead propagates the `None`. After this is merged, the behaviour will be such that the COM interface is emulated. If a method fails, an exception won't be raised and instead `None` will be returned.

Reviewed-on: https://git.alt.lu/alt/pywellcad/pulls/24
Co-authored-by: Matthew Schubert <matthew.schubert@alt.lu>
Co-committed-by: Matthew Schubert <matthew.schubert@alt.lu>