Clarify behaviour of left/right position setting (#73)
The documentation for `left_position`/`right_position` for classes `Depth`, `Title` and `Log` needs to be updated to clarify the behaviour when the left position is set to a greater value than the right (or the right to a value less than the left).

This should allow for removal of a few failing tests, too.

Reviewed-on: https://git.alt.lu/alt/pywellcad/pulls/73
Co-authored-by: Matthew Schubert <matthew.schubert@alt.lu>
Co-committed-by: Matthew Schubert <matthew.schubert@alt.lu>